Separate InfoPaneHours intl a separate component
